BUG#26796118 : instance rejoins GR group after removeInstance() and restart
After removing an instance from a cluster, using the removeInstance()
function, the instances silently rejoins the Group Replication (GR) group
after it is restarted. This happens because the GR plugin automatically start
GR when the server is restarted (group_replication_start_on_boot=ON by
default).

This patch fixes this issue by setting group_replication_start_on_boot to
OFF, using SET PERSIT for server version >= 8.0.2. For server versions
without the SET PERSIST feature a warning is issued to notify the users
that the must persist this setting on the configuration file manually to
avoid this issue.

In addition, this patch also fixes a few issues in the test suite:
- improve validation of the TMPDIR environment variable, required for unit
  tests;
- fix sandbox self deployment features, to allow tests to be execute
  individually;
- improve testutil.destroySandbox() to provide an option not to stop the
  server, since it might be already stopped (avoiding errors trying to kill
  it);

Appropriate unit tests were added for regression.

+++ unittest/test_utils/mod_testutils.cc
@@ -92,10 +92,11 @@ Testutils::Testutils(const std::string &sandbox_dir, bool dummy_mode,
     std::cerr << "tetutils using dummy sandboxes\n";
 
   expose("deploySandbox", &Testutils::deploy_sandbox, "port", "rootpass");
-  expose("destroySandbox", &Testutils::destroy_sandbox, "port");
+  expose("destroySandbox", &Testutils::destroy_sandbox, "port", "?quiet_kill",
+         false);
   expose("startSandbox", &Testutils::start_sandbox, "port");
   expose("stopSandbox", &Testutils::stop_sandbox, "port", "rootpass");
-  expose("killSandbox", &Testutils::kill_sandbox, "port");
+  expose("killSandbox", &Testutils::kill_sandbox, "port", "?quiet", false);
   expose("restartSandbox", &Testutils::restart_sandbox, "port", "rootpass");
 
   expose("snapshotSandboxConf", &Testutils::snapshot_sandbox_conf, "port");
@@ -340,22 +341,15 @@ void Testutils::end_snapshot_sandbox_error_log(int port) {
 void Testutils::deploy_sandbox(int port, const std::string &rootpass) {
   mysqlshdk::db::replay::No_replay dont_record;
   if (!_dummy_sandboxes) {
-    if (true) {
-      // Sandbox from a boilerplate
-      if (!_boilerplate_rootpass.empty() && _boilerplate_rootpass == rootpass &&
-          !_expected_boilerplate_version.empty()) {
+    // Sandbox from a boilerplate
+    if (!_boilerplate_rootpass.empty() && _boilerplate_rootpass == rootpass &&
+        !_expected_boilerplate_version.empty()) {
+      if (!deploy_sandbox_from_boilerplate(port)) {
+        prepare_sandbox_boilerplate(rootpass, port);
         if (!deploy_sandbox_from_boilerplate(port)) {
-          prepare_sandbox_boilerplate(rootpass);
-          if (!deploy_sandbox_from_boilerplate(port)) {
-            std::cerr << "Unable to deploy boilerplate sandbox\n";
-            abort();
-          }
+          std::cerr << "Unable to deploy boilerplate sandbox\n";
+          abort();
         }
-      } else {
-        prepare_sandbox_boilerplate(rootpass);
-        _boilerplate_rootpass = rootpass;
-
-        deploy_sandbox_from_boilerplate(port);
       }
     } else {
       // Sandbox from scratch
@@ -394,9 +388,9 @@ void Testutils::deploy_sandbox(int port, const std::string &rootpass) {
   None Testutils::destroy_sandbox(int port);
 #endif
 ///@}
-void Testutils::destroy_sandbox(int port) {
+void Testutils::destroy_sandbox(int port, bool quiet_kill) {
   mysqlshdk::db::replay::No_replay dont_record;
-  kill_sandbox(port);
+  kill_sandbox(port, quiet_kill);
 #ifdef _WIN32
   // Make config file (and backups) readable in case it was made RO by some test
   std::string dirname = shcore::path::dirname(get_sandbox_conf_path(port));
@@ -565,11 +559,12 @@ void Testutils::restart_sandbox(int port, const std::string &rootpass) {
   None Testutils::kill_sandbox(int port);
 #endif
 ///@}
-void Testutils::kill_sandbox(int port) {
+void Testutils::kill_sandbox(int port, bool quiet) {
   if (!_dummy_sandboxes) {
     shcore::Value::Array_type_ref errors;
     _mp->kill_sandbox(port, _sandbox_dir, &errors);
-    if (errors && !errors->empty())
+    // Only output errors to stderr if quiet mode is disabled (default).
+    if (!quiet && errors && !errors->empty())
       std::cerr << "During kill of " << port << ": "
                 << shcore::Value(errors).descr() << "\n";
     wait_sandbox_dead(port);
@@ -912,7 +907,8 @@ std::string Testutils::fetch_captured_stderr() {
   return _fetch_stderr();
 }
 
-void Testutils::prepare_sandbox_boilerplate(const std::string &rootpass) {
+void Testutils::prepare_sandbox_boilerplate(const std::string &rootpass,
+                                            int port) {
   if (g_test_trace_scripts)
     std::cerr << "Preparing sandbox boilerplate...\n";
 
@@ -929,8 +925,6 @@ void Testutils::prepare_sandbox_boilerplate(const std::string &rootpass) {
 
   // Create a sandbox, shut it down and then keep a copy of its basedir
   // to be reused for future deployments
-  int port =
-      _default_sandbox_ports.empty() ? 3300 : _default_sandbox_ports.front();
 
   shcore::Value::Array_type_ref errors;
   shcore::Value mycnf_options = shcore::Value::new_array();
